# edge-agent/openapi.yaml
openapi: 3.0.3
info:
  title: 3dxAgent Edge-Agent API
  version: 3.2.0
  description: >
    REST-API des 3dxAgent Fusions-Kernels. Bietet EKF-Zustand, Historie,
    Szenarien-Konfiguration, ICP-Map-Merging, die v2.0-Datenpipeline, die
    Aura-Endpunkte (RTI/Heatmap), die CT45P-Triangulation sowie den
    Datenexport (GeoJSON/KML/JSON)
    (docs/AURA.md, docs/TRIANGULATION.md, docs/SERVICE_WORKER.md).
servers:
  - url: http://localhost:8080
paths:
  /api/v1/health:
    get:
      summary: Health-Check
      responses:
        '200':
          description: Status des Agenten
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string }
                  mode: { type: string }
  /api/v1/agent/state:
    get:
      summary: Aktueller 6-DOF-EKF-Zustand
      responses:
        '200':
          description: EKF-Zustand
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EkfState'
  /api/v1/agent/config:
    post:
      summary: Szenario starten/konfigurieren
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScenarioConfig'
      responses:
        '200':
          description: Bestätigung
  /api/v1/agent/history:
    get:
      summary: 3D-Transformations-Historie
      parameters:
        - name: device_id
          in: query
          schema: { type: string, default: CT45P-01 }
        - name: limit
          in: query
          schema: { type: integer, default: 100 }
      responses:
        '200':
          description: Liste von Transformationen
  /api/v1/agent/merge:
    post:
      summary: ICP-Map-Merging mehrerer CT45P
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MergeRequest'
      responses:
        '200':
          description: Ergebnis des Mergings
        '400':
          description: Weniger als 2 Geräte
  /api/v1/pipeline/run:
    post:
      summary: v2.0-Datenpipeline ausführen
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PipelineRequest'
      responses:
        '200':
          description: Pipeline-Ergebnis
  /api/v1/aura/rti:
    post:
      summary: Aura RTI-Rekonstruktion (Messlinien → Voxel-Dämpfungsfeld)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuraRtiRequest'
      responses:
        '200':
          description: Voxel-Feld, Peaks, Rekonstruktions-Metadaten
        '400':
          description: Weniger als 2 Messlinien
  /api/v1/aura/heatmap:
    post:
      summary: Aura RF-Samples → extrudierte Heatmap-Zellen
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuraHeatmapRequest'
      responses:
        '200':
          description: Extrudierte Zellen (Höhe ∝ Signalstärke)
        '400':
          description: Keine Samples übergeben
  /api/v1/triangulation/solve:
    post:
      summary: Trilateration — Anker + Distanzen → Position
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TriangulationRequest'
      responses:
        '200':
          description: Positionsschätzung (inkl. Residuum, Sigma, Konfidenz)
        '400':
          description: 'Weniger als 3 gültige Anker (3D: 4)'
  /api/v1/export:
    post:
      summary: Datenexport — Annotationen/Punkte → GeoJSON/KML/JSON
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExportRequest'
      responses:
        '200':
          description: Konvertierter Export-Inhalt
        '400':
          description: Unbekanntes Format oder keine Daten
  /api/v1/network/topology:
    post:
      summary: Network3D — Topologie-Ingest (Upsert) + Broadcast
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TopologyRequest'
      responses:
        '200':
          description: Node-/Edge-Zähler + Snapshot-Index
    get:
      summary: Aktuelle Topologie
      responses:
        '200':
          description: Nodes + Edges der Live-Topologie
  /api/v1/network/simulate:
    post:
      summary: Network3D — What-If-Failover-Simulation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SimulationRequest'
      responses:
        '200':
          description: Betroffene/reroutete/unerreichbare Flows
        '404':
          description: Node nicht in der Topologie
  /api/v1/network/history:
    get:
      summary: Network3D — Time-Machine-Snapshot-Replay
      parameters:
        - name: index
          in: query
          schema: { type: integer }
        - name: limit
          in: query
          schema: { type: integer, default: 100 }
      responses:
        '200':
          description: Snapshots (oder ein einzelner Snapshot bei index)
  /api/v1/network/devices:
    get:
      summary: Live-Netzwerk — Geräte des Trackers
      responses:
        '200':
          description: Bekannte Geräte (Change-/Anomalie-Erkennung)
components:
  schemas:
    EkfState:
      type: object
      properties:
        x: { type: number }
        y: { type: number }
        z: { type: number }
        vx: { type: number }
        vy: { type: number }
        vz: { type: number }
        covariance:
          type: array
          items:
            type: array
            items: { type: number }
        kalman_gain_lidar: { type: number }
        mode: { type: string }
    ScenarioConfig:
      type: object
      required: [type]
      properties:
        type: { type: string, enum: [tactical, evacuation, architecture, temp, research] }
        params: { type: object, additionalProperties: true }
    MergeRequest:
      type: object
      required: [device_ids]
      properties:
        device_ids:
          type: array
          items: { type: string }
        reference: { type: string }
    PipelineRequest:
      type: object
      required: [device_id]
      properties:
        device_id: { type: string }
        points:
          type: array
          items: { type: number }
        metadata:
          type: object
          additionalProperties: true
    AuraRtiRequest:
      type: object
      properties:
        device_id: { type: string, default: CT45P-01 }
        bounds_min:
          type: array
          items: { type: number }
        bounds_max:
          type: array
          items: { type: number }
        voxel_size: { type: number, default: 0.5 }
        ellipse_width: { type: number, default: 0.05 }
        regularization: { type: number, default: 0.1 }
        method: { type: string, enum: [tikhonov, backprojection], default: tikhonov }
        links:
          type: array
          items:
            type: object
            required: [tx, rx, attenuation_db]
            properties:
              tx:
                type: array
                items: { type: number }
              rx:
                type: array
                items: { type: number }
              attenuation_db: { type: number }
    AuraHeatmapRequest:
      type: object
      properties:
        device_id: { type: string, default: CT45P-01 }
        samples:
          type: array
          items:
            type: object
            properties:
              x: { type: number }
              y: { type: number }
              z: { type: number }
              dbm: { type: number }
              frequency_hz: { type: number }
        cell_size_m: { type: number, default: 1.0 }
    TriangulationRequest:
      type: object
      properties:
        anchors:
          type: array
          items:
            type: object
            properties:
              id: { type: string }
              x: { type: number }
              y: { type: number }
              z: { type: number }
        distances:
          type: object
          additionalProperties: { type: number }
        uncertainties:
          type: object
          additionalProperties: { type: number }
        use_z: { type: boolean, default: false }
    ExportRequest:
      type: object
      properties:
        format: { type: string, enum: [geojson, kml, json], default: geojson }
        annotations:
          type: array
          items:
            type: object
            properties:
              id: { type: string }
              title: { type: string }
              description: { type: string }
              lon: { type: number }
              lat: { type: number }
              z: { type: number }
        points:
          type: array
          items:
            type: array
            items: { type: number }
        device_id: { type: string, default: CT45P-01 }
    TopologyRequest:
      type: object
      properties:
        nodes:
          type: array
          items:
            type: object
            properties:
              id: { type: string }
              type: { type: string, default: host }
              name: { type: string }
              status: { type: string, default: ok }
              x: { type: number }
              y: { type: number }
              z: { type: number }
        edges:
          type: array
          items:
            type: object
            properties:
              id: { type: string }
              source: { type: string }
              target: { type: string }
              bandwidth_mbps: { type: number }
              latency_ms: { type: number }
              utilization: { type: number }
    SimulationRequest:
      type: object
      required: [node_id]
      properties:
        node_id: { type: string }
        flows:
          type: array
          items:
            type: object
            properties:
              id: { type: string }
              source: { type: string }
              target: { type: string }